Add files via upload
Here we focus on the pure categorical datasets. The first issue is imputation Accuracy. As shown in the figure(from left to right are HR, ZO, SB datasets) ,  ATMI consistently achieves the highest accuracy. Surprisingly, the simple MS method performs well on the ZO and SB datasets, especially at a 50% missing rate, comparable to MF and KNNMI. This suggests that when the missing rate is high and data relationships are severely damaged, simple methods can still be effective.